Refactor Telegram integration: update TelegramService to use new API, enhance TelegramApp rendering, and introduce TelegramChatRepository

+++ src/main/kotlin/space/davids_digital/kiri/model/telegram/TelegramPhotoSize.kt
@@ -19,5 +19,5 @@ data class TelegramPhotoSize (
     /**
      * File size in bytes.
      */
-    val fileSize: Int? = null,
+    val fileSize: Long? = null,
 )
